First successful pass through linting & type checking for newly factored `utility` module.
- Still needs type annotations.
