+++ tox.ini
@@ -76,7 +76,7 @@ skip_install = true
 deps =
     flake8
 commands =
-    flake8 --ignore=E501,E272,E241,E222,E221 src/
+    flake8 --ignore=E501,E272,E241,E222,E221,F401,F403,W291 src/
 
 [testenv:type-check]
 description = Run type checking using `mypy`.
